requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
discord.py==2.1.1
exceptiongroup==1.1.1
Flask==2.2.3
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.5
multidict==6.0.4
nodeenv==1.8.0
prisma==0.8.2
pydantic==1.10.8
python-dotenv==0.21.1
redis==4.5.4
requests==2.28.2
six==1.16.0
sniffio==1.3.0
spotipy==2.23.0
table2ascii==0.5.0
tomlkit==0.11.8
twitchio==2.5.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
